refactor: split large test file into multiple files
